+++ crates/cervo-core/src/inferer.rs
@@ -208,6 +208,24 @@ pub trait InfererExt: Inferer + Sized {
 
 impl<T> InfererExt for T where T: Inferer + Sized {}
 
+impl Inferer for Box<dyn Inferer + Send> {
+    fn select_batch_size(&self, max_count: usize) -> usize {
+        self.as_ref().select_batch_size(max_count)
+    }
+
+    fn infer_raw(&self, batch: ScratchPadView<'_>) -> Result<(), anyhow::Error> {
+        self.as_ref().infer_raw(batch)
+    }
+
+    fn input_shapes(&self) -> &[(String, Vec<usize>)] {
+        self.as_ref().input_shapes()
+    }
+
+    fn output_shapes(&self) -> &[(String, Vec<usize>)] {
+        self.as_ref().output_shapes()
+    }
+}
+
 impl Inferer for Box<dyn Inferer> {
     fn select_batch_size(&self, max_count: usize) -> usize {
         self.as_ref().select_batch_size(max_count)
