+++ crates/cervo-runtime/Cargo.toml
@@ -17,3 +17,6 @@ readme = "../../README.md"
 cervo-core = { version = "0.2.1-alpha.0", path = "../cervo-core"}
 thiserror = "1.0"
 anyhow = "1.0.41"
+
+[dev-dependencies]
+cervo-asset = { version = "0.2.1-alpha.0", path = "../cervo-asset"}
